requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.3.0
langchain-google-genai>=3.0.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
pydantic>=2.8.0
rank-bm25>=0.2.2
pytest>=8.0.0
httpx>=0.27.0
